[fix] переименовал названия пакетов
