+++ src/main/java/units/abstractUnits/UnitAttackingWithWeapons.java
@@ -1,4 +1,4 @@
-package units.Abstract;
+package units.abstractUnits;
 
 import units.UnitsTypes;
