+++ src/main/java/weapons/Weapon.java
@@ -1,4 +1,4 @@
-package Weapons;
+package weapons;
 
 public class Weapon {
 }
